requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.7
django-jazzmin==2.6.0
environ==1.0
idna==3.4
python-dotenv==0.21.0
pytz==2022.7
requests==2.28.1
sqlparse==0.4.3
tzdata==2022.7
urllib3==1.26.13
xlwt==1.3.0
qrcode
user-agents
folium
drf-spectacular